third_party/blink/web_tests/external/wpt/css/css-values/random-computed.tentative-expected.txt

This is a testharness.js-based test.
Found 65 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Property scale value 'random(1, 11)'
  assert_true: 'random(1, 11)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(--foo, 2, 12)'
  assert_true: 'random(--foo, 2, 12)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(--foo per-element, 3, 13)'
  assert_true: 'random(--foo per-element, 3, 13)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(per-element --foo, 4, 14)'
  assert_true: 'random(per-element --foo, 4, 14)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(0, 10, by 5)'
  assert_true: 'random(0, 10, by 5)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(--foo, 10, 20, by 5)'
  assert_true: 'random(--foo, 10, 20, by 5)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(--foo per-element, 20, 30, by 5)'
  assert_true: 'random(--foo per-element, 20, 30, by 5)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(per-element --foo, 30, 40, by 5)'
  assert_true: 'random(per-element --foo, 30, 40, by 5)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(100, 10)'
  assert_true: 'random(100, 10)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(-10, -100)'
  assert_true: 'random(-10, -100)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(-100, -10)'
  assert_true: 'random(-100, -10)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(40, 50, by -5)'
  assert_true: 'random(40, 50, by -5)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(5 * 1, 30 / 2)'
  assert_true: 'random(5 * 1, 30 / 2)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(2 * random(6, 16))'
  assert_true: 'calc(2 * random(6, 16))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(NaN, 100)'
  assert_true: 'random(NaN, 100)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(10, NaN)'
  assert_true: 'random(10, NaN)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(NaN, NaN)'
  assert_true: 'random(NaN, NaN)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(NaN, 100, by 10)'
  assert_true: 'random(NaN, 100, by 10)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(10, NaN, by 10)'
  assert_true: 'random(10, NaN, by 10)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(NaN, NaN, by 10)'
  assert_true: 'random(NaN, NaN, by 10)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(NaN, 100, by NaN)'
  assert_true: 'random(NaN, 100, by NaN)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(10, NaN, by NaN)'
  assert_true: 'random(10, NaN, by NaN)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(NaN, NaN, by NaN)'
  assert_true: 'random(NaN, NaN, by NaN)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(10, 100, by NaN)'
  assert_true: 'random(10, 100, by NaN)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(NaN, 100))'
  assert_true: 'calc(10 + random(NaN, 100))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(10, NaN))'
  assert_true: 'calc(10 + random(10, NaN))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(NaN, NaN))'
  assert_true: 'calc(10 + random(NaN, NaN))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(NaN, 100, by 10))'
  assert_true: 'calc(10 + random(NaN, 100, by 10))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(10, NaN, by 10))'
  assert_true: 'calc(10 + random(10, NaN, by 10))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(NaN, NaN, by 10))'
  assert_true: 'calc(10 + random(NaN, NaN, by 10))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(NaN, 100, by NaN))'
  assert_true: 'calc(10 + random(NaN, 100, by NaN))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(10, NaN, by NaN))'
  assert_true: 'calc(10 + random(10, NaN, by NaN))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(NaN, NaN, by NaN))'
  assert_true: 'calc(10 + random(NaN, NaN, by NaN))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(10, 100, by NaN))'
  assert_true: 'calc(10 + random(10, 100, by NaN))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(infinity, 100)'
  assert_true: 'random(infinity, 100)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(10, infinity)'
  assert_true: 'random(10, infinity)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(infinity, infinity)'
  assert_true: 'random(infinity, infinity)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(infinity, 100, by 10)'
  assert_true: 'random(infinity, 100, by 10)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(10, infinity, by 10)'
  assert_true: 'random(10, infinity, by 10)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(infinity, infinity, by 10)'
  assert_true: 'random(infinity, infinity, by 10)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(infinity, 100, by infinity)'
  assert_true: 'random(infinity, 100, by infinity)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(10, infinity, by infinity)'
  assert_true: 'random(10, infinity, by infinity)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(infinity, infinity, by infinity)'
  assert_true: 'random(infinity, infinity, by infinity)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(infinity, 100))'
  assert_true: 'calc(10 + random(infinity, 100))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(10, infinity))'
  assert_true: 'calc(10 + random(10, infinity))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(infinity, infinity))'
  assert_true: 'calc(10 + random(infinity, infinity))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(infinity, 100, by 10))'
  assert_true: 'calc(10 + random(infinity, 100, by 10))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(10, infinity, by 10))'
  assert_true: 'calc(10 + random(10, infinity, by 10))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(infinity, infinity, by 10))'
  assert_true: 'calc(10 + random(infinity, infinity, by 10))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(infinity, 100, by infinity))'
  assert_true: 'calc(10 + random(infinity, 100, by infinity))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(10, infinity, by infinity))'
  assert_true: 'calc(10 + random(10, infinity, by infinity))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(infinity, infinity, by infinity))'
  assert_true: 'calc(10 + random(infinity, infinity, by infinity))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(10, 100, by infinity)'
  assert_true: 'random(10, 100, by infinity)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'random(10, 100, by -infinity)'
  assert_true: 'random(10, 100, by -infinity)' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(10, 100, by infinity))'
  assert_true: 'calc(10 + random(10, 100, by infinity))' is a supported value for scale. expected true got false
[FAIL] Property scale value 'calc(10 + random(10, 100, by -infinity))'
  assert_true: 'calc(10 + random(10, 100, by -infinity))' is a supported value for scale. expected true got false
[FAIL] Property scale value on pseudo element '::before' 'random(7, 17)'
  assert_greater_than_equal: random(7, 17) expected a number greater than or equal to 7 but got NaN
[FAIL] Property scale value on pseudo element '::before' 'random(--bar, 8, 18)'
  assert_greater_than_equal: random(--bar, 8, 18) expected a number greater than or equal to 8 but got NaN
[FAIL] Property scale value on pseudo element '::before' 'random(per-element, 9, 19)'
  assert_greater_than_equal: random(per-element, 9, 19) expected a number greater than or equal to 9 but got NaN
[FAIL] Property scale value on pseudo element '::before' 'random(per-element --foo, 10, 20)'
  assert_greater_than_equal: random(per-element --foo, 10, 20) expected a number greater than or equal to 10 but got NaN
[FAIL] Uses of a style rule for two properies on an element with identical uses of random(...).
  assert_equals: expected "0px" but got "784px"
[FAIL] Uses of a style rule for two properies on an element with identical uses of random(--identifier, ...).
  assert_equals: expected "0px" but got "784px"
[FAIL] Uses of a style rule for two properies on an element with identical uses of random(per-elemnent, ...).
  assert_equals: expected "0px" but got "784px"
[FAIL] Uses of a style rule for two properies on an element with identical uses of random(per-elemnent --identifier, ...).
  assert_equals: expected "0px" but got "784px"
[FAIL] Uses of a style rule on multiple elements with 'random(per-element, ...)'
  assert_false: expected false got true
Harness: the test ran to completion.
